Migrate from PodIdentity to WorkloadIdentity
Fixed the ACR Pull issue - was able to remove several steps from the demo setup how-to
Added some outputs to the Bicep template to simplify data retrieval and command execution in the how-to section
Removed the "ConnectionString" setting in appsettings.json for the OrderGenerator.  It was a placeholder which was overridden locally by environment variable, but when running in k8s caused an auth failure.  If the connection string is present, it doesn't try to use RBAC auth from the Pod or Workload Identity.
Added an option in the how-to to use ACR Build tasks to build the container images (for those without Docker locally).
Updated the README.md